Split viewsets into separate files (#101)
