+++ backend/fms_core/viewsets/__init__.py
@@ -1,0 +1,29 @@
+from .container import ContainerViewSet
+from .container_kind import ContainerKindViewSet
+from .experiment_run import ExperimentRunViewSet
+from .group import GroupViewSet
+from .individual import IndividualViewSet
+from .process_measurement import ProcessMeasurementViewSet
+from .protocol import ProtocolViewSet
+from .query import QueryViewSet
+from .revision import RevisionViewSet
+from .sample import SampleViewSet
+from .sample_kind import SampleKindViewSet
+from .user import UserViewSet
+from .version import VersionViewSet
+
+__all__ = [
+    "ContainerViewSet",
+    "ContainerKindViewSet",
+    "ExperimentRunViewSet",
+    "GroupViewSet",
+    "IndividualViewSet",
+    "ProcessMeasurementViewSet",
+    "ProtocolViewSet",
+    "QueryViewSet",
+    "RevisionViewSet",
+    "SampleViewSet",
+    "SampleKindViewSet",
+    "UserViewSet",
+    "VersionViewSet",
+]
